began overlying structure
added very rough structure to tie together the major components of any
pipeline, included at the very beginning of the script. not checked. i
don't even think it has a closed {}. also the rest of the script was not
tabbed correctly to reflect that it is included within the original
structure. might even be backwards. idk.